# packages/api-client/runtime.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GridForge Edge Runtime",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/demo": {
      "get": {
        "summary": "Demo Read",
        "operationId": "demo_read_api_v1_demo_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Demo Step",
        "operationId": "demo_step_api_v1_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoStep"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/identity/bootstrap": {
      "post": {
        "summary": "Identity Bootstrap",
        "operationId": "identity_bootstrap_api_v1_identity_bootstrap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/identity/logout": {
      "post": {
        "summary": "Identity Logout",
        "operationId": "identity_logout_api_v1_identity_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/system/bundle": {
      "get": {
        "summary": "Diagnostic Bundle",
        "operationId": "diagnostic_bundle_api_v1_system_bundle_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Diagnostic Bundle Api V1 System Bundle Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registry": {
      "get": {
        "summary": "Registry",
        "operationId": "registry_api_v1_registry_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrySnapshot"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save Registry",
        "operationId": "save_registry_api_v1_registry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryRecord"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DemoEvidence": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "normal",
              "bad",
              "peak",
              "propose",
              "approve",
              "verify",
              "explain",
              "disconnect",
              "reconnect"
            ],
            "title": "Action"
          },
          "simulated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Simulated At"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["action", "simulated_at", "detail"],
        "title": "DemoEvidence"
      },
      "DemoSnapshot": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "SIMULATION",
            "title": "Mode",
            "default": "SIMULATION"
          },
          "clock": {
            "type": "string",
            "const": "ACCELERATED_ISOLATED",
            "title": "Clock",
            "default": "ACCELERATED_ISOLATED"
          },
          "ephemeral": {
            "type": "boolean",
            "const": true,
            "title": "Ephemeral",
            "default": true
          },
          "edge_id": {
            "type": "string",
            "format": "uuid",
            "title": "Edge Id"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "simulated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Simulated At"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "normal",
                  "bad",
                  "peak",
                  "propose",
                  "approve",
                  "verify",
                  "explain",
                  "disconnect",
                  "reconnect"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/DemoEvidence"
            },
            "type": "array",
            "title": "Steps"
          },
          "telemetry_rows": {
            "type": "integer",
            "title": "Telemetry Rows"
          },
          "demand_kw": {
            "type": "number",
            "title": "Demand Kw"
          },
          "forecast_status": {
            "type": "string",
            "title": "Forecast Status"
          },
          "risk": {
            "type": "string",
            "title": "Risk"
          },
          "run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizationRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "command": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DispatchCommand"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verification"
              },
              {
                "type": "null"
              }
            ]
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CopilotAnswer"
              },
              {
                "type": "null"
              }
            ]
          },
          "sync": {
            "$ref": "#/components/schemas/SyncSnapshot"
          },
          "audit_integrity_ok": {
            "type": "boolean",
            "title": "Audit Integrity Ok"
          },
          "audit_head": {
            "type": "string",
            "title": "Audit Head"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "edge_id",
          "org_id",
          "facility_id",
          "simulated_at",
          "next_action",
          "steps",
          "telemetry_rows",
          "demand_kw",
          "forecast_status",
          "risk",
          "run",
          "command",
          "verification",
          "explanation",
          "sync",
          "audit_integrity_ok",
          "audit_head"
        ],
        "title": "DemoSnapshot"
      },
      "DemoStep": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "normal",
              "bad",
              "peak",
              "propose",
              "approve",
              "verify",
              "explain",
              "disconnect",
              "reconnect"
            ],
            "title": "Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["request_id", "action"],
        "title": "DemoStep"
      },
...
      "IdentityStatus": {
        "properties": {
          "setup_required": {
            "type": "boolean",
            "title": "Setup Required"
          },
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalUser"
              },
              {
                "type": "null"
              }
            ]
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permissions"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "SIMULATION"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "setup_required",
          "authenticated",
          "user",
          "permissions",
          "expires_at"
        ],
        "title": "IdentityStatus"
      },
...
      "Login": {
        "properties": {
          "username": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_.-]{3,50}$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["username", "password"],
        "title": "Login"
      },
...
      "RegistryRecord": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "entity": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Organization"
              },
              {
                "$ref": "#/components/schemas/Facility"
              },
              {
                "$ref": "#/components/schemas/ProductionLine"
              },
              {
                "$ref": "#/components/schemas/Asset"
              },
              {
                "$ref": "#/components/schemas/Device"
              },
              {
                "$ref": "#/components/schemas/services__registry__contracts__Metric"
              },
              {
                "$ref": "#/components/schemas/SignalMapping"
              }
            ],
            "title": "Entity",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "asset": "#/components/schemas/Asset",
                "device": "#/components/schemas/Device",
                "facility": "#/components/schemas/Facility",
                "line": "#/components/schemas/ProductionLine",
                "mapping": "#/components/schemas/SignalMapping",
                "metric": "#/components/schemas/services__registry__contracts__Metric",
                "organization": "#/components/schemas/Organization"
              }
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["revision", "entity"],
        "title": "RegistryRecord"
      },
      "RegistrySnapshot": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1",
            "title": "Schema Version",
            "default": "1"
          },
          "mode": {
            "type": "string",
            "const": "SIMULATION",
            "title": "Mode",
            "default": "SIMULATION"
          },
          "org_id": {
            "type": "string",
            "format": "uuid",
            "title": "Org Id"
          },
          "facility_id": {
            "type": "string",
            "format": "uuid",
            "title": "Facility Id"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/RegistryRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "connectors": {
            "items": {
              "$ref": "#/components/schemas/ConnectorHealth"
            },
            "type": "array",
            "title": "Connectors"
          },
          "flexibility_available": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flexibility Available"
          },
          "pending_configuration_events": {
            "type": "integer",
            "title": "Pending Configuration Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "org_id",
          "facility_id",
          "records",
          "connectors",
          "flexibility_available",
          "pending_configuration_events"
        ],
        "title": "RegistrySnapshot"
      },
      "RegistryWrite": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "entity": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Organization"
              },
              {
                "$ref": "#/components/schemas/Facility"
              },
              {
                "$ref": "#/components/schemas/ProductionLine"
              },
              {
                "$ref": "#/components/schemas/Asset"
              },
              {
                "$ref": "#/components/schemas/Device"
              },
              {
                "$ref": "#/components/schemas/Metric-Input"
              },
              {
                "$ref": "#/components/schemas/SignalMapping"
              }
            ],
            "title": "Entity",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "asset": "#/components/schemas/Asset",
                "device": "#/components/schemas/Device",
                "facility": "#/components/schemas/Facility",
                "line": "#/components/schemas/ProductionLine",
                "mapping": "#/components/schemas/SignalMapping",
                "metric": "#/components/schemas/Metric-Input",
                "organization": "#/components/schemas/Organization"
              }
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["expected_revision", "request_id", "entity"],
        "title": "RegistryWrite"
      },
...
      "RuntimeErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Details"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["code", "message", "details", "request_id"],
        "title": "RuntimeErrorResponse"
      },
...
    },
    "securitySchemes": {
      "LocalSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "LocalSession": []
    }
  ]
}